adding media upload

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *    //Only allowing the index age and the show pae
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -42,13 +52,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
           'title' => 'required',
-          'body' => 'required'
+          'body' => 'required',
+          'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File upload
+        if($request->hasFile('cover_image')){
+          // Get filename with the extension
+          $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+          // Get just the filename
+          $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+          // Get just the extension
+          $extension = $request->file('cover_image')->getClientOriginalExtension();
+          // FileName to Store
+          $fileNameToStore = $filename.'_'.time().'.'.$extension;
+          // Upload the Image
+          $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else{
+          $fileNameToStore = 'noimage.jpg';
+        }
+
         // Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Blog Post Created');
@@ -62,9 +91,10 @@ class PostsController extends Controller
      */
     public function show($id)
     {
-        // return Post::find($id);
         $post = Post::find($id);
-        // return view('posts.show');
+        if(!$post){
+          return redirect('/posts')->with('error', 'Post not found');
+        }
         return view('posts.show')->with('post', $post);
     }
 
@@ -76,8 +106,18 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        //
+
         $post = Post::find($id);
+        if(!$post){
+          return redirect('/posts')->with('error', 'Post not found');
+        }
+
+        //Checking for the correct user
+
+        if(auth()->user()->id !== $post->user_id){
+          return redirect('/posts')->with('error', 'Unauthorized');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -112,6 +152,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Checking for the correct user
+        if(auth()->user()->id !== $post->user_id){
+          return redirect('/posts')->with('error', 'Unauthorized');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Blog Post Deleted');
     }
